requirements.txt
feedparser
requests
beautifulsoup4
python-dateutil
google-generativeai
